+++ packages/core/tests/graph/edges.test.ts
@@ -1,6 +1,11 @@
 import { describe, expect, it, vi } from 'vitest';
 import type { IPlugin } from '@codegraphy-dev/plugin-api';
 import type { IProjectedConnection } from '../../src/analysis/projectedConnection';
+import {
+  createGraphEdgeId,
+  getGraphEdgeIdSuffix,
+  replaceGraphEdgeIdEndpoints,
+} from '../../src/graph/edgeIdentity';
 import { buildWorkspaceGraphEdges } from '../../src/graph/edges';
 
 function createPlugin(id: string): IPlugin {
@@ -292,5 +297,26 @@ describe('core/graph/edges', () => {
         ],
       },
     ]);
+  });
+
+  it('creates edge ids with optional type and variant suffixes', () => {
+    expect(createGraphEdgeId({ from: 'a.ts', to: 'b.ts', kind: 'import' })).toBe('a.ts->b.ts#import');
+    expect(createGraphEdgeId({
+      from: 'a.ts',
+      to: 'b.ts',
+      kind: 'reference',
+      type: 'value',
+      variant: 'dynamic',
+    })).toBe('a.ts->b.ts#reference:value~dynamic');
+  });
+
+  it('reads and preserves edge id suffixes when endpoints change', () => {
+    expect(getGraphEdgeIdSuffix('a.ts->b.ts#import:value~dynamic', 'import')).toBe('#import:value~dynamic');
+    expect(getGraphEdgeIdSuffix('legacy-edge-id', 'reference')).toBe('#reference');
+    expect(replaceGraphEdgeIdEndpoints(
+      { id: 'a.ts->b.ts#import:value~dynamic', kind: 'import' },
+      'src/new-a.ts',
+      'src/new-b.ts',
+    )).toBe('src/new-a.ts->src/new-b.ts#import:value~dynamic');
   });
 });
