+++ packages/core/tests/plugins/packageManifest.test.ts
@@ -1,8 +1,21 @@
+import type { IPlugin } from '@codegraphy-dev/plugin-api';
 import { describe, expect, it } from 'vitest';
 
 import {
   parseCodeGraphyPluginPackageManifest,
 } from '../../src';
+import { assertPluginApiCompatibility } from '../../src/plugins/compatibility';
+import { compareSemver, parseSemver } from '../../src/plugins/semverParts';
+
+function plugin(apiVersion: unknown): IPlugin {
+  return {
+    id: 'codegraphy.test',
+    name: 'Test',
+    version: '1.0.0',
+    apiVersion: apiVersion as string,
+    supportedExtensions: ['.test'],
+  };
+}
 
 describe('CodeGraphy plugin package manifest', () => {
   it('accepts plugin metadata from package.json without importing runtime code', () => {
@@ -94,5 +107,32 @@ describe('CodeGraphy plugin package manifest', () => {
         apiVersion: '',
       },
     })).toBeNull();
+  });
+
+  it('accepts compatible plugin API ranges and rejects unsupported versions', () => {
+    expect(() => assertPluginApiCompatibility(plugin('2'))).not.toThrow();
+    expect(() => assertPluginApiCompatibility(plugin('^2.0.0'))).not.toThrow();
+    expect(() => assertPluginApiCompatibility(plugin(undefined))).toThrow(
+      "Plugin 'codegraphy.test' must declare a string apiVersion",
+    );
+    expect(() => assertPluginApiCompatibility(plugin('^3.0.0'))).toThrow(
+      "Plugin 'codegraphy.test' targets unsupported CodeGraphy Plugin API '^3.0.0'.",
+    );
+  });
+
+  it('parses exact semantic versions and rejects non-exact versions', () => {
+    expect(parseSemver('  2.10.3  ')).toEqual({ major: 2, minor: 10, patch: 3 });
+    expect(parseSemver('12.10.34')).toEqual({ major: 12, minor: 10, patch: 34 });
+    expect(parseSemver('2.10')).toBeUndefined();
+    expect(parseSemver('^2.10.3')).toBeUndefined();
+    expect(parseSemver('2.10.3-beta')).toBeUndefined();
+  });
+
+  it('compares semantic versions by major, then minor, then patch', () => {
+    expect(compareSemver({ major: 2, minor: 0, patch: 0 }, { major: 1, minor: 9, patch: 9 })).toBeGreaterThan(0);
+    expect(compareSemver({ major: 1, minor: 9, patch: 9 }, { major: 2, minor: 0, patch: 0 })).toBeLessThan(0);
+    expect(compareSemver({ major: 2, minor: 1, patch: 0 }, { major: 2, minor: 3, patch: 0 })).toBeLessThan(0);
+    expect(compareSemver({ major: 2, minor: 1, patch: 5 }, { major: 2, minor: 1, patch: 3 })).toBeGreaterThan(0);
+    expect(compareSemver({ major: 2, minor: 1, patch: 5 }, { major: 2, minor: 1, patch: 5 })).toBe(0);
   });
 });
